backend-agent/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# LangChain dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-community>=0.0.1
langgraph>=0.0.1
openai>=1.10.0

# HTTP client
httpx>=0.26.0

# MCP Server (local dependency)
# Note: Install mcp-server dependencies separately

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
